Drop old /help, add basic ratelimir sys (not finished)
